jdoc 1
- all teh logic for sanitizing
- the transformer for @tag prompts
- the parsing and prompting logic for methods

still required
- function, that prefixes the comment's lines with the ' * ' thing and
  '/**' and '*/' bits
- the parsing and prompting for
   - class and interface
   - static final
   - constructors
- the function, that recognizes which type of line we're on

Can I just say, that the declarative-rewrite is genius? This whole
commit added like 194 lines, with comments and all. Add what's still
missing and a little more docs and you're at 300 or so. The original
jdoc solution clocks in at 900 lines -- it's literally half of the old
code!